refactor: correccion utils
