+++ src/app/infrastructure/ui/hooks/useCars/useCars.ts
@@ -2,24 +2,8 @@ import { useEffect, useState } from 'react'
 import { createCarService } from '../../../factories/car.factory'
 import { createLocationService } from '../../../factories/location.factory'
 import ICar from '../../../interfaces/car.interface'
+import { getDateFromParam } from '../../../lib/util'
 import { useAuthentication } from '../../common/contexts/AuthenticationContext/AuthenticationContext'
-
-const getDateFromParam = (dateParam: string) => {
-  const arrayString = []
-  let index = 0
-  const CHUNK_LENGTH = 2
-  const MAX_LENGTH_DATE = 3
-
-  while (true) {
-    arrayString.push(dateParam.substr(index, CHUNK_LENGTH))
-    index += CHUNK_LENGTH
-    if (arrayString.length === MAX_LENGTH_DATE) break
-  }
-
-  arrayString.reverse()
-
-  return `20${arrayString.join('-')}`
-}
 
 interface ISearchParams {
   originCityCode: string
